Support upgrading GCC

+++ _emerge.sh
@@ -10,6 +10,13 @@ configs() {
 	done
 }
 
+if emerge -pv gcc | grep -q NS; then
+	old_gcc=$(gcc-config -S $(gcc-config -c) | cut -d\  -f2)
+	emerge -1u gcc && gcc-config 2 || exit 1
+	env-update && . profile
+	emerge -C =sys-devel/gcc-$old_gcc && emerge -1 libtool || exit 1
+fi
+
 emerge -e git world || exit 1
 emerge -c || exit 1
 rm -fr ../var/portage/distfiles/*
